publish @living-dictionaries/parts to enable usage inside firebase functions, finish updating move of partsOfSpeech and semanticDomains
