+++ packages/functions/package.json
@@ -1,5 +1,6 @@
 {
-  "name": "@ld/functions",
+  "name": "@living-dictionaries/functions",
+  "version": "0.0.1",
   "scripts": {
     "build": "tsc",
     "build:watch": "tsc --watch",
@@ -26,7 +27,6 @@
   "devDependencies": {
     "@living-dictionaries/types": "^1.0.0",
     "@living-dictionaries/parts": "^0.0.1",
-    "@living-dictionaries/scripts": "^0.0.1",
     "@firebase/testing": "^0.20.11",
     "fetch-as": "^1.1.0",
     "firebase-functions-test": "^0.2.0",
