+++ packages/functions/src/algolia/prepareDataForIndex.ts
