+++ packages/functions/src/export/interfaceExplanations.ts
@@ -1,172 +1,3 @@
-import { ISemanticDomain } from '@living-dictionaries/parts';
-
-export const semanticDomains: ISemanticDomain[] = [
-  { key: '1', name: 'Universe and the natural world' },
-  { key: '1.1', name: 'Sky, weather and climate' },
-  { key: '1.2', name: 'Earth, geology and landscape' },
-  { key: '1.3', name: 'Water, bodies of water and the water cycle' },
-  { key: '1.4', name: 'Plants, trees and other vegetation' },
-  { key: '1.5', name: 'Animals' },
-  { key: '1.6', name: 'Colors' },
-  { key: '1.7', name: 'Birds' },
-  { key: '1.8', name: 'Fish and marine creatures' },
-  { key: '1.9', name: 'Insects and small creatures' },
-  { key: '1.10', name: 'Reptiles and amphibians' },
-  { key: '2', name: 'Humans, Social Relations and Organization' },
-  { key: '2.1', name: 'Body parts' },
-  { key: '2.2', name: 'Body functions' },
-  { key: '2.3', name: 'Senses and perception' },
-  { key: '2.4', name: 'Health, well-being and sickness' },
-  { key: '2.5', name: 'Life stages and rites of passage' },
-  { key: '2.6', name: 'Kinship and relationships' },
-  { key: '2.7', name: 'Social activities and interactions' },
-  { key: '2.8', name: 'Government and Law' },
-  { key: '3', name: 'Language and Communication' },
-  { key: '3.1', name: 'Mental actions' },
-  { key: '3.2', name: 'Emotions' },
-  { key: '3.3', name: 'Education and learning' },
-  { key: '4', name: 'Names' },
-  { key: '4.1', name: 'Names of people and clans' },
-  { key: '4.2', name: 'Place names' },
-  { key: '4.3', name: 'Sacred names' },
-  { key: '5', name: 'Daily life' },
-  { key: '5.1', name: 'Household' },
-  { key: '5.2', name: 'Agriculture' },
-  { key: '5.3', name: 'Food, cooking and foodways' },
-  { key: '5.4', name: 'Clothing and adornment' },
-  { key: '5.5', name: 'Cleanliness and hygiene' },
-  { key: '5.6', name: 'Sleep' },
-  { key: '5.7', name: 'Work and occupations' },
-  { key: '5.8', name: 'Animal husbandry' },
-  { key: '5.9', name: 'Tools and weapons' },
-  { key: '5.10', name: 'Finance and Business' },
-  { key: '5.11', name: 'Hunting and Fishing' },
-  { key: '5.12', name: 'Art, crafts and textiles' },
-  { key: '5.13', name: 'Construction' },
-  { key: '5.14', name: 'Sports and games' },
-  { key: '5.15', name: 'Transportation' },
-  { key: '6', name: 'Physical Actions and States' },
-  { key: '6.1', name: 'Posture' },
-  { key: '6.2', name: 'Motion' },
-  { key: '6.3', name: 'Possession' },
-  { key: '6.4', name: 'Physical Impact' },
-  { key: '6.5', name: 'States' },
-  { key: '7', name: 'Quantities and Quantifiers' },
-  { key: '7.1', name: 'Numerals' },
-  { key: '7.2', name: 'Sizes and dimensions' },
-  { key: '7.3', name: 'Units of Measure' },
-  { key: '7.4', name: 'Quantifiers' },
-  { key: '8', name: 'Time' },
-  { key: '9', name: 'Location and Directions' },
-  { key: '10', name: 'Grammar' },
-  { key: '10.1', name: 'Case Markers / Adpositions ' },
-  { key: '10.2', name: 'TAM' },
-  { key: '10.3', name: 'Polarity' },
-  { key: '10.4', name: 'Coordinators, Subordinators, Relativizers, Quotatives' },
-  { key: '10.5', name: 'Discourse Markers' },
-  { key: '10.6', name: 'Interjections' },
-  { key: '10.7', name: 'Ideophones, Expressives & Onomatopoeia' },
-  { key: '10.8', name: 'Classifiers' },
-  { key: '10.9', name: 'Pro-forms' },
-];
-
-export const partsOfSpeech = {
-  n: 'noun',
-  pro: 'pronoun',
-  v: 'verb',
-  adj: 'adjective',
-  adv: 'adverb',
-  prep: 'preposition',
-  conj: 'conjunction',
-  det: 'determiner',
-  int: 'interjection',
-  no: 'number',
-  dem: 'demonstrative',
-  'n pl': 'plural noun',
-  phr: 'phrase',
-  p: 'particle',
-  neg: 'negative',
-  quan: 'quantifier',
-  rel: 'relative',
-  'v.aux': 'auxiliary verb',
-  'dem.a': 'animate demonstrative',
-  'dem.i': 'inanimate demonstrative',
-  'dem.loc': 'locative demonstrative',
-  dim: 'diminutive',
-  na: 'animate noun',
-  ni: 'inanimate noun',
-  pl: 'plural',
-  pn: 'prenoun',
-  pv: 'preverb',
-  'pn/v': 'prenoun / preverb',
-  'pro.a': 'animate pronoun',
-  'pro.i': 'inanimate pronoun',
-  redup: 'reduplication',
-  refl: 'reflexive',
-  vai: 'animate intransitive verb',
-  'vai/ii': 'animate / inanimate intransitive verb',
-  vii: 'inanimate intransitive verb',
-  vta: 'transitive animate verb',
-  vti: 'transitive inanimate verb',
-  imp: 'imperative/command marker',
-  pref: 'prefix',
-  punc: 'punctuation',
-  suff: 'suffix',
-  vt: 'transitive verb',
-  vi: 'intransitive verb',
-  comp: 'complement(izer)',
-  cpd: 'compound',
-  pple: 'participle',
-  sent: 'sentence',
-  'pr.n': 'proper noun',
-  np: 'noun phrase',
-  vp: 'verb phrase',
-  ap: 'adjective phrase',
-  'adv.p': 'adverb phrase',
-  q: 'question marker',
-  ger: 'gerund',
-  art: 'article',
-  post: 'postposition',
-  'e.w': 'entry word / discourse particle',
-  'm.w': 'measure word',
-  msd: 'masdar',
-  idf: 'idafa',
-  'v.prdg': 'verbal paradigm',
-  acr: 'acronym',
-  cntr: 'contraction',
-  num: 'numeral',
-  'act.v': 'active verb',
-  'stat.v': 'stative verb',
-  'caus.v': 'causative verb',
-  'v.suff': 'verb suffix',
-  'n.suff': 'noun suffix',
-  'n.pref': 'noun prefix',
-  'v.pref': 'verb prefix',
-  ideo: 'ideophone (onomatoepia / expressives)',
-  clf: 'classifier',
-  tns: 'tense marker',
-  asp: 'aspect marker',
-  mod: 'mood marker',
-  'n.cls': 'noun class marker',
-  quot: 'quotative',
-  adp: 'adposition',
-  def: 'definite article',
-  indf: 'indefinite article',
-  indfpro: 'indefinite pronoun',
-  pers: 'personal pronoun',
-  poss: 'possessive pronoun',
-  reflpro: 'reflexive pronoun',
-  relpro: 'relative pronoun',
-  vd: 'ditransitive verb',
-  'interr.pro': 'interrogative pronoun',
-  cop: 'copula',
-  'imp.verb': 'imperative verb',
-  'phr.v': 'phrasal verb',
-  'inal.n': 'inalienable noun',
-  'rel.n': 'relational noun',
-  'v.n': 'verbal noun',
-};
-
 export const entryInterface = {
   id:
     'Unique Entry ID - use this for links to the entry on Living Dictionaries, as well as to compare new imports with previous data',
