+++ packages/parts/package.json
@@ -5,14 +5,10 @@
   "type": "module",
   "svelte": "src/lib/index.ts",
   "main": "src/lib/index.ts",
-  "publishConfig": {
-    "main": "index.js",
-    "typings": "index.d.ts"
-  },
   "scripts": {
     "dev": "svelte-kit dev --port 3042",
     "build": "svelte-kit build",
-    "package": "svelte-kit package",
+    "package": "svelte-kit package && node update-packaged-index-path.js",
     "preview": "svelte-kit preview",
     "prepare": "svelte-kit sync",
     "check": "svelte-check --tsconfig ./tsconfig.json --threshold warning --diagnostic-sources js,svelte",
@@ -31,6 +27,7 @@
     "mapbox-gl": "^2.8.1",
     "mdsvex": "^0.10.5",
     "recordrtc": "^5.6.2",
+    "replace-in-file": "^6.3.2",
     "svelte": "^3.47.0",
     "svelte-check": "^2.7.0",
     "svelte-deep-wind-preprocess": "^0.0.7",
@@ -38,6 +35,7 @@
     "svelte-preprocess": "^4.10.6",
     "svelte-store-router": "^2.0.10",
     "svelte-windicss-preprocess": "~4.1.0",
+    "svelte2tsx": "^0.5.10",
     "tslib": "^2.3.1",
     "typescript": "~4.6.2",
     "vitest": "^0.10.0",
@@ -47,8 +45,7 @@
     "type": "git",
     "url": "git+https://github.com/livingtongues/living-dictionaries.git"
   },
-  "author": "Living Tongues",
-  "license": "MIT",
+  "author": "Living Tongues Institute for Endangered Languages",
   "bugs": {
     "url": "https://github.com/livingtongues/living-dictionaries/issues"
   },
