--- packages/scripts/index.ts
@@ -1,1 +1,0 @@
-export { prepareDataForIndex } from './algolia/prepareDataForIndex';
