+++ packages/scripts/package.json
@@ -14,6 +14,7 @@
     "test": "vitest -u"
   },
   "devDependencies": {
+    "@living-dictionaries/functions": "workspace:^0.0.1",
     "@living-dictionaries/types": "^1.0.0",
     "@types/node": "^17.0.23",
     "algoliasearch": "^4.11.0",
