+++ packages/site/src/lib/components/table/cells/SelectPOS.svelte
@@ -6,7 +6,7 @@
   export let value = '',
     canEdit = false;
 
-  import { partsOfSpeech } from '$lib/mappings/parts-of-speech';
+  import { partsOfSpeech } from '@living-dictionaries/parts';
 
   $: writeInPOS = value && !partsOfSpeech.find((part) => part.enAbbrev === value);
